Review changes #624
